move input loading function
