+++ day-two/src/solution-pt1.ts
@@ -1,16 +1,4 @@
-import inputLoader from "./inputLoader";
-
-export const loadData = async (): Promise<string[]> => {
-  try {
-    const data: Buffer = await inputLoader();
-    const dataInString: string = data.toString();
-    const parsed = dataInString.split("\r\n");
-
-    return parsed;
-  } catch (exception) {
-    throw new Error(`Failed to load input data. ${exception.message}`);
-  }
-};
+import inputData from "./inputLoader";
 
 export const scanBoxId = () => {};
